+++ UC4-chaincode/src/main/java/de/upb/cs/uc4/chaincode/contract/operation/OperationContractUtil.java
@@ -3,10 +3,13 @@ package de.upb.cs.uc4.chaincode.contract.operation;
 import de.upb.cs.uc4.chaincode.contract.ContractUtil;
 import de.upb.cs.uc4.chaincode.contract.group.GroupContractUtil;
 import de.upb.cs.uc4.chaincode.exceptions.serializable.LedgerAccessError;
+import de.upb.cs.uc4.chaincode.exceptions.serializable.parameter.MissingTransactionError;
+import de.upb.cs.uc4.chaincode.helper.AccessManager;
 import de.upb.cs.uc4.chaincode.model.ApprovalList;
 import de.upb.cs.uc4.chaincode.model.OperationData;
+import de.upb.cs.uc4.chaincode.model.OperationDataState;
+import de.upb.cs.uc4.chaincode.model.TransactionInfo;
 import de.upb.cs.uc4.chaincode.model.errors.DetailedError;
-import de.upb.cs.uc4.chaincode.model.errors.GenericError;
 import de.upb.cs.uc4.chaincode.model.errors.InvalidParameter;
 import org.hyperledger.fabric.contract.Context;
 import org.hyperledger.fabric.shim.ChaincodeStub;
@@ -14,14 +17,13 @@ import org.hyperledger.fabric.shim.ChaincodeStub;
 import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
-import java.util.ArrayList;
 import java.util.Base64;
 import java.util.List;
 import java.util.stream.Collectors;
 
 public class OperationContractUtil extends ContractUtil {
     private static final String HASH_DELIMITER = ":";
-    private static final GroupContractUtil groupContractUtil = new GroupContractUtil();
+    private static final GroupContractUtil gUtil = new GroupContractUtil();
 
     public OperationContractUtil() {
         keyPrefix = "operation:";
@@ -33,18 +35,34 @@ public class OperationContractUtil extends ContractUtil {
         return message;
     }
 
-    public String getSystemDetailedRejectionMessage(DetailedError error) {
-        return "The Transaction failed with an error of type: " + error.getType();
+    public OperationData approveOperation(Context ctx, OperationData operationData) throws MissingTransactionError {
+        String clientId = this.getEnrollmentIdFromClientId(ctx.getClientIdentity().getId());
+        List<String> clientGroups = new GroupContractUtil().getGroupNamesForUser(ctx.getStub(), clientId);
+
+        ApprovalList existingApprovals = operationData.getExistingApprovals().addUsersItem(clientId).addGroupsItems(clientGroups);
+        TransactionInfo info = operationData.getTransactionInfo();
+        ApprovalList requiredApprovals = AccessManager.getRequiredApprovals(info.getContractName(), info.getTransactionName(), info.getParameters());
+
+        ApprovalList missingApprovals = OperationContractUtil.getMissingApprovalList(requiredApprovals, existingApprovals);
+        return operationData.lastModifiedTimestamp(this.getTimestamp(ctx.getStub()))
+                .existingApprovals(existingApprovals)
+                .missingApprovals(missingApprovals);
     }
 
-    public String getSystemGenericRejectionMessage(GenericError error) {
-        return "The Transaction failed with an error of type: " + error.getType();
-    }
-
-    public List<OperationData> getOperations(ChaincodeStub stub, final String existingEnrollmentId, final String missingEnrollmentId, final String initiatorEnrollmentId, String state) {
-        List<String> groupsForUserMissingApproval = new GroupContractUtil().getGroupNamesForUser(stub, missingEnrollmentId);
+    public List<OperationData> getOperations(
+            ChaincodeStub stub,
+            final List<String> operationIds,
+            final String existingEnrollmentId,
+            final String missingEnrollmentId,
+            final String initiatorEnrollmentId,
+            final String involvedEnrollmentId,
+            final List<String> states) {
+        List<String> groupsForUserMissingApproval = gUtil.getGroupNamesForUser(stub, missingEnrollmentId);
+        List<String> groupsForUserInvolved = gUtil.getGroupNamesForUser(stub, involvedEnrollmentId);
 
         return this.getAllStates(stub, OperationData.class).stream()
+                .filter(item -> operationIds.isEmpty() ||
+                        operationIds.contains(item.getOperationId()))
                 .filter(item -> existingEnrollmentId.isEmpty() ||
                         item.getExistingApprovals().getUsers().contains(existingEnrollmentId))
                 .filter(item -> missingEnrollmentId.isEmpty() ||
@@ -52,8 +70,14 @@ public class OperationContractUtil extends ContractUtil {
                         item.getMissingApprovals().getGroups().stream().anyMatch(groupsForUserMissingApproval::contains))
                 .filter(item -> initiatorEnrollmentId.isEmpty() ||
                         item.getInitiator().equals(initiatorEnrollmentId))
-                .filter(item -> state.isEmpty() ||
-                        item.getState().toString().equals(state)).collect(Collectors.toList());
+                .filter(item -> involvedEnrollmentId.isEmpty() ||
+                        item.getExistingApprovals().getUsers().contains(involvedEnrollmentId) ||
+                        item.getMissingApprovals().getUsers().contains(involvedEnrollmentId) ||
+                        item.getMissingApprovals().getGroups().stream().anyMatch(groupsForUserInvolved::contains) ||
+                        item.getInitiator().equals(initiatorEnrollmentId))
+                .filter(item -> states.isEmpty() ||
+                        states.contains(item.getState().toString()))
+                .collect(Collectors.toList());
     }
 
     public static boolean covers(ApprovalList requiredApprovals, ApprovalList existingApprovals) {
@@ -74,17 +98,6 @@ public class OperationContractUtil extends ContractUtil {
         return new String(Base64.getUrlEncoder().encode(bytes));
     }
 
-    public ArrayList<InvalidParameter> getErrorForInput(String contractName, String transactionName) {
-        ArrayList<InvalidParameter> invalidParams = new ArrayList<>();
-        if (valueUnset(contractName)) {
-            invalidParams.add(getEmptyInvalidParameter("contractName"));
-        }
-        if (valueUnset(transactionName)) {
-            invalidParams.add(getEmptyInvalidParameter("transactionName"));
-        }
-        return invalidParams;
-    }
-
     public DetailedError getContractUnprocessableError(String contractName) {
         return getUnprocessableEntityError(new InvalidParameter()
                 .name("contractName")
@@ -95,5 +108,32 @@ public class OperationContractUtil extends ContractUtil {
         return getUnprocessableEntityError(new InvalidParameter()
                 .name("transactionName")
                 .reason("The given transaction \"" + transactionName + "\" does not exist"));
+    }
+
+    public DetailedError getEmptyContractNameError() {
+        return getUnprocessableEntityError(getEmptyInvalidParameter("contractName"));
+    }
+
+    public DetailedError getEmptyTransactionNameError() {
+        return getUnprocessableEntityError(getEmptyInvalidParameter("transactionName"));
+    }
+
+    public OperationData getOrInitializeOperationData(Context ctx, String initiator, String contractName, String transactionName, String params) throws NoSuchAlgorithmException {
+        String key = OperationContractUtil.getDraftKey(contractName, transactionName, params);
+        OperationData operationData;
+        String timeStamp = getTimestamp(ctx.getStub());
+        try {
+            operationData = getState(ctx.getStub(), key, OperationData.class);
+        } catch (LedgerAccessError ledgerAccessError) {
+            operationData = new OperationData()
+                    .initiator(initiator)
+                    .operationId(key)
+                    .initiatedTimestamp(timeStamp)
+                    .transactionInfo(new TransactionInfo().contractName(contractName).transactionName(transactionName).parameters(params))
+                    .state(OperationDataState.PENDING)
+                    .reason("");
+
+        }
+        return operationData;
     }
 }